# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ePortfolio API",
        "description": "API para el sistema de portfolio electrónico de estudiantes",
        "contact": {
            "name": "ePortfolio Support",
            "email": "admin@eportfolio.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        },
        {
            "url": "http://eportfolio.test/api/v1",
            "description": "Local development server"
        },
        {
            "url": "https://eportfolio.cifpcarlos3.es/api/v1",
            "description": "Production server"
        },
        {
            "url": "https://eportfolio.juezlti.eu/api/v1",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/roles": {
            "get": {
                "tags": [
                    "Rol"
                ],
                "summary": "List all rols",
                "description": "Retrieve a paginated list of rols",
                "operationId": "219aa99f0ea40b14fbb5f3b9dadc6ad1",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Rol"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rol"
                ],
                "summary": "Create a new rol",
                "description": "Create a new rol resource",
                "operationId": "1a4844f88745b12f75ecda89bd20fc13",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreRolRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Rol"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/roles/{id}": {
            "get": {
                "tags": [
                    "Rol"
                ],
                "summary": "Show a specific rol",
                "description": "Retrieve a specific rol by ID",
                "operationId": "2d60f8015c5031ab0253f9fcab7e9adf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Rol"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Rol"
                ],
                "summary": "Update a specific rol",
                "description": "Update a specific rol by ID",
                "operationId": "610553774a54e7dadc84dfeaa45d043e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRolRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resource updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Rol"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rol"
                ],
                "summary": "Delete a specific rol",
                "description": "Delete a specific rol by ID",
                "operationId": "5bb912c54b58cb39916f46cb9b73200b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rol eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreRolRequest": {
                "title": "Store Role Request",
                "description": "Datos requeridos para crear un Rol",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre único del rol",
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "description": "Descripción del rol",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UpdateRolRequest": {
                "title": "Update Role Request",
                "description": "Datos para actualizar un Rol",
                "properties": {
                    "name": {
                        "description": "Nombre único del rol",
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "description": "Descripción del rol",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Rol": {
                "title": "Rol",
                "description": "Modelo de Rol",
                "properties": {
                    "id": {
                        "description": "ID único",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del rol",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descripción del rol",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}